x86_64 fixes
Commit of intptr_t %)
Since xfig uses pointer args to pass integer values there were some
warnings about different size of int and void * on 64-bit system.
To fix this problem I changed int to intptr_t in some places.
Such places include:
- local variable types
- static global variable types
- one or two global exported variable types
- rarely explicit conversion type specification
28 of 80 warnings of that type are still there because their correction
requires change of globally spread variable types and data structures.

+++ xfig/e_movept.c
@@ -133,7 +133,7 @@ init_move_point(F_line *obj, int type, int x, int y, F_point *p, F_point *q)
       case O_ELLIPSE:
 	force_noanglegeom();
 	/* dirty trick - arcpoint_num is stored in p */
-	movedpoint_num = (int) p;
+	movedpoint_num = (intptr_t) p;
 	cur_e = (F_ellipse *) obj;
 	if (!init_ellipsepointmoving()) /* selected center, ignore */
 	    return False;
@@ -141,7 +141,7 @@ init_move_point(F_line *obj, int type, int x, int y, F_point *p, F_point *q)
       case O_ARC:
 	force_noanglegeom();
 	/* dirty trick - arcpoint_num is stored in p */
-	movedpoint_num = (int) p;
+	movedpoint_num = (intptr_t) p;
 	cur_a = (F_arc *) obj;
 	init_arcpointmoving();
 	break;
